Fix duplicate params

+++ data/typing/numpy.lib.stride_tricks.py
@@ -1,4 +1,4 @@
 
 
-def as_strided(x: numpy.ndarray, shape: tuple[(Union[(numpy.int64, int)], ...)], strides: tuple[(Union[(numpy.int64, int)], ...)]):
+def as_strided(x: numpy.ndarray, shape: tuple[(Union[(numpy.int64, int)], ...)], strides: tuple[(Union[(int, numpy.int64)], ...)]):
     'usage.skimage: 11'
